Move NERO States To Separate File
