+++ Core/Inc/state_machine.h
@@ -3,18 +3,9 @@
 
 #include "cmsis_os.h"
 #include "fault.h"
+#include "nero_state_machine.h"
 
 typedef enum { BOOT, READY, DRIVING, FAULTED, MAX_FUNC_STATES } func_state_t;
-
-typedef enum {
-	NOT_DRIVING,
-	REVERSE,
-	PIT,
-	EFFICIENCY,
-	PERFORMANCE,
-   	NERO_ONLY,
-	MAX_DRIVE_STATES
-} drive_state_t;
 
 extern osThreadId_t sm_director_handle;
 extern const osThreadAttr_t sm_director_attributes;
@@ -32,26 +23,6 @@ func_state_t get_func_state();
  * Will return negative if functional state is not DRIVING
  */
 int queue_drive_state(drive_state_t new_state);
-
-/*
- * Attempts to increment the nero index and handles any exceptions 
-*/
-void increment_nero_index();
-
-/*
- * Attempts to decrement the nero index and handles any exceptions
-*/
-void decrement_nero_index();
-
-/*
- * Tells the statemachine to track up and down movements
-*/
-void set_home_mode();
-
-/*
- * Tells NERO to select the current index if in home mode
-*/
-void select_nero_index();
 
 
 /*
